Ships now sync correctly

+++ Assets/Scripts/Game/Common/Settings/NetworkSettings.cs
@@ -5,6 +5,6 @@ namespace Game.Common.Settings
     [Serializable]
     public struct NetworkSettings
     {
-        public float updateInterval;
+        public ulong updateInterval;
     }
 }
